Add dtype=float for Nifti1Pair image creation.
See https://github.com/nipy/nibabel/issues/1089

np.array using int64 for arrays created with all integers. These
arrays cannot be passed to directly to Nifti1Pair constructor
anymore without explicit dtype argument specifying that the user
wants a int64 image. I added a dtype=float specification to the
np.array construction to get rid of this error.

+++ tests/image/test_image.py
@@ -71,7 +71,7 @@ def spatial_image() -> SpatialImage:
                                  [[0, 0, 0, 0],
                                   [0, 0, 0, 0],
                                   [0, 0, 0, 0],
-                                  [0, 0, 0, 0]]]]).reshape(4, 4, 4, 1),
+                                  [0, 0, 0, 0]]]], dtype=float).reshape(4, 4, 4, 1),
                       np.eye(4))
